refactor: move /internal to /pkg/internal
